feat: Sandbox scene
The sandbox scene is a general purpose scene. We will use it to experiment different setups and game mechanics there. This first scene is just an experiment to mimic the basic scene setup from Kenneys.nl, learning the basics to import models and set some simple world environment and lights that work out of the box.